requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiomysql==0.2.0
anyio==3.7.0
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.2
attrs==23.1.0
audioread==3.0.1
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
boto3==1.26.93
botocore==1.29.93
Brlapi==0.8.3
bs4==0.0.1
cachetools==5.3.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
comm==0.1.2
command-not-found==0.3
contourpy==1.0.7
control==0.9.4
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
deep-sort-realtime==1.3.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.109.2
fasteners==0.14.1
fastjsonschema==2.17.1
filelock==3.10.0
filterpy==1.4.5
fire==0.5.0
Flask==2.3.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
flatbuffers==23.3.3
fonttools==4.39.0
fqdn==1.5.1
fsspec==2024.3.1
future==0.18.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.17.1
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-cloud-speech==2.25.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==2.0.2
grpcio==1.62.0
grpcio-status==1.62.0
gyp==0.1
h11==0.14.0
h5py==3.8.0
httplib2==0.20.2
huggingface-hub==0.13.2
idna==2.10
imageio==2.27.0
importlib-metadata==4.6.4
ipykernel==6.21.2
ipython==8.11.0
ipython-genutils==0.2.0
iso8601==1.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
keras==2.12.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==16.0.0
librosa==0.10.1
llvmlite==0.42.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
meson==0.61.2
mistune==2.0.5
ml-dtypes==0.0.4
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.3
mss==9.0.1
munkres==1.1.4
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nashpy==0.0.36
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.59.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.72
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.1.0
pooch==1.8.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pybboxes==0.1.6
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.6
PyGetWindow==0.0.9
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyMySQL==1.0.3
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
python3-xlib==0.15
pytweening==1.0.7
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.0
rdkit==2022.9.5
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
s3transfer==0.6.0
sahi==0.11.12
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.2
sentry-sdk==1.19.1
serial==0.0.97
shapely==2.0.1
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
SQLAlchemy==2.0.13
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.36.3
sympy==1.12
systemd-python==234
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
terminado==0.17.1
terminaltables==3.1.10
texttable==1.6.4
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.3.21
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.2.0
typing_extensions==4.9.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.0.69
unattended-upgrades==0.1
uri-template==1.2.0
urllib3==1.26.15
usb-creator==0.3.7
uvicorn==0.25.0
vboxapi==1.0
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==2.3.4
wrapt==1.14.1
WTForms==3.0.1
xdg==5
xkit==0.0.0
yapf==0.32.0
yolov5==7.0.9
yolov8==0.0.1
zipp==1.0.0
